Move placeholders to a dedicated namespace
