+++ osu.Game/Online/Leaderboards/RetrievalFailurePlaceholder.cs
@@ -6,6 +6,7 @@ using osu.Framework.Graphics;
 using osu.Framework.Graphics.Sprites;
 using osu.Framework.Input.Events;
 using osu.Game.Graphics.Containers;
+using osu.Game.Online.Placeholders;
 using osuTK;
 
 namespace osu.Game.Online.Leaderboards
