+++ osu.Game/Online/Placeholders/MessagePlaceholder.cs
@@ -4,7 +4,7 @@
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Sprites;
 
-namespace osu.Game.Online.Leaderboards
+namespace osu.Game.Online.Placeholders
 {
     public class MessagePlaceholder : Placeholder
     {
